# Backend - nginx/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Реализация RestAPI на Go с использование Swagger для документации",
        "title": "School 21 ApiFromSpec docs",
        "termsOfService": "http://swagger.io/terms",
        "contact": {
            "name": "desmonsu"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/suppliers": {
            "get": {
                "description": "Возвращает массив поставщиков. Если товары не найдены, возвращает пустой массив.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Получение всех поставщиков",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apifromspec_internal_pkg.GetSupplierResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "fail to fetch suppliers",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавление поставщика в базу данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Добавление поставщика",
                "parameters": [
                    {
                        "description": "Данные для добавления поставщика",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.AddSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "supplier added successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "wrong input format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "supplier already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/suppliers/{id}": {
            "get": {
                "description": "Возвращает найденного поставщика",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Получение поставщика по id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.GetSupplierResponse"
                        }
                    },
                    "400": {
                        "description": "can not parse id from URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "supplier not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "fail to fetch supplier",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Возвращает информацию об успешности операции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Удаляет поставщика из базы данных",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID поставщика для удаления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "supplier deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "wrong input format or can not parse id from URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "supplier not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to delete supplier",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/suppliers/{id}/address": {
            "patch": {
                "description": "Заменяет адрес поставщика в базе данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Смена адреса для поставщика",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Параметры нового адреса поставщика",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.AddressChangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "address changed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "wrong input format or validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "id not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to process address",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "apifromspec_internal_pkg.AddSupplierRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Moscow"
                },
                "country": {
                    "type": "string",
                    "example": "Russia"
                },
                "name": {
                    "type": "string",
                    "example": "Tech Wholesale"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+70001112233"
                },
                "street": {
                    "type": "string",
                    "example": "Tverskaya"
                }
            }
        },
        "apifromspec_internal_pkg.AddressChangeRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Saint Petersburg"
                },
                "country": {
                    "type": "string",
                    "example": "Russia"
                },
                "street": {
                    "type": "string",
                    "example": "Nevskiy Prospect"
                }
            }
        },
...
        "apifromspec_internal_pkg.GetSupplierResponse": {
            "type": "object",
            "properties": {
                "address_id": {
                    "type": "string",
                    "example": "660f8400-e29b-41d4-a716-446655440011"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Tech Wholesale"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+70001112233"
                }
            }
        },
...
    }
}